+++ src/meta.jl
@@ -4,7 +4,7 @@ information about the most recent resolve.
 """
 
 # increment whenever the metadata format changes
-const META_VERSION = 13
+const META_VERSION = 14
 
 @kwdef mutable struct Meta
     timestamp::Float64
@@ -12,6 +12,7 @@ const META_VERSION = 13
     load_path::Vector{String}
     extra_path::Vector{String}
     version::VersionNumber
+    backend::Symbol
     packages::Vector{PkgSpec}
     channels::Vector{ChannelSpec}
     pip_packages::Vector{PipPkgSpec}
@@ -26,6 +27,7 @@ function read_meta(io::IO)
             load_path = read_meta(io, Vector{String}),
             extra_path = read_meta(io, Vector{String}),
             version = read_meta(io, VersionNumber),
+            backend = read_meta(io, Symbol),
             packages = read_meta(io, Vector{PkgSpec}),
             channels = read_meta(io, Vector{ChannelSpec}),
             pip_packages = read_meta(io, Vector{PipPkgSpec}),
@@ -42,6 +44,9 @@ function read_meta(io::IO, ::Type{String})
         error("unexpected end of meta file")
     end
     String(bytes)
+end
+function read_meta(io::IO, ::Type{Symbol})
+    Symbol(read_meta(io, String))
 end
 function read_meta(io::IO, ::Type{Vector{T}}) where {T}
     len = read(io, Int)
@@ -81,6 +86,7 @@ function write_meta(io::IO, meta::Meta)
     write_meta(io, meta.load_path)
     write_meta(io, meta.extra_path)
     write_meta(io, meta.version)
+    write_meta(io, meta.backend)
     write_meta(io, meta.packages)
     write_meta(io, meta.channels)
     write_meta(io, meta.pip_packages)
@@ -92,6 +98,9 @@ end
 function write_meta(io::IO, x::String)
     write(io, convert(Int, sizeof(x)))
     write(io, x)
+end
+function write_meta(io::IO, x::Symbol)
+    write_meta(io, String(x))
 end
 function write_meta(io::IO, x::Vector)
     write(io, convert(Int, length(x)))