+++ src/spec.jl
@@ -87,6 +87,22 @@ function specstr(x::PkgSpec)
     string(x.name, suffix)
 end
 
+function pixispec(x::PkgSpec)
+    spec = Dict{String,Any}()
+    spec["version"] = x.version == "" ? "*" : x.version
+    if x.build != ""
+        spec["build"] = x.build
+    end
+    if x.channel != ""
+        spec["channel"] = x.channel
+    end
+    if length(spec) == 1
+        spec["version"]
+    else
+        spec
+    end
+end
+
 struct ChannelSpec
     name::String
     function ChannelSpec(name)
@@ -189,4 +205,38 @@ function specstr(x::PipPkgSpec)
         push!(parts, " ", x.version)
     end
     return join(parts)
+end
+
+function pixispec(x::PipPkgSpec)
+    spec = Dict{String,Any}()
+    if startswith(x.version, "@")
+        url = strip(x.version[2:end])
+        if startswith(url, "git+")
+            url = url[5:end]
+            if (m = match(r"^(.*?)@([^/@]*)$", url); m !== nothing)
+                spec["git"] = m.captures[1]
+                rev = m.captures[2]
+                if (m = match(r"^([^#]*)#(.*)$", rev); m !== nothing)
+                    # spec["tag"] = m.captures[1]
+                    spec["rev"] = m.captures[2]
+                else
+                    spec["rev"] = rev
+                end
+            else
+                spec["git"] = url
+            end
+        else
+            spec["url"] = url
+        end
+    else
+        spec["version"] = x.version == "" ? "*" : x.version
+    end
+    if !isempty(x.extras)
+        spec["extras"] = x.extras
+    end
+    if length(spec) == 1 && haskey(spec, "version")
+        spec["version"]
+    else
+        spec
+    end
 end