+++ test/setup.jl
@@ -6,14 +6,16 @@ ENV["JULIA_CONDAPKG_VERBOSITY"] = "0"
 
 status() = sprint(io -> CondaPkg.status(io = io))
 
-const backend = get(ENV, "JULIA_CONDAPKG_BACKEND", "MicroMamba")
+const backend = get(ENV, "JULIA_CONDAPKG_BACKEND", "Pixi")
 
 const isnull = backend == "Null"
+const ispixi = backend == "SystemPixi" || backend == "Pixi"
 
 # reset the package state (so tests are independent of the order they are run)
 rm(CondaPkg.cur_deps_file(), force = true)
 CondaPkg.STATE.backend = :NotSet
 CondaPkg.STATE.condaexe = ""
+CondaPkg.STATE.pixiexe = ""
 CondaPkg.STATE.resolved = false
 CondaPkg.STATE.load_path = String[]
 CondaPkg.STATE.meta_dir = ""