+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## Unreleased
-* Add `SystemPixi` backend to allow using [Pixi](https://pixi.sh/latest/) to install packages.
+* Add `Pixi` and `SystemPixi` backends to allow using [Pixi](https://pixi.sh/latest/) to install packages.
 
 ## 0.2.24 (2024-11-08)
 * Add `pip_backend` preference to choose between `pip` and `uv`.